+++ src/translations/en.ts
@@ -737,6 +737,11 @@ export const en = {
   publisher: {
     selectbutton: 'Select other {{type}}',
     nodataset: 'No DataSet referenced',
+    smv: {
+      commsetting: 'Communication Settings (SMV)',
+      noconnectedap: 'No connection to SubNetwork',
+      smvopts: 'Optional Fields',
+    },
   },
   userinfo: {
     loggedInAs: 'Logged in as {{name}}',